Working config
Still need to add in websockets code
