+++ data/index.html
@@ -79,6 +79,7 @@
             <button id="pilotSwapButton4" onclick="pilotSwap('team4', this.id)">Pilot Swap</button>
         </div>
     </div>
+    <div>Version 1.1.03</div>
     <script src="script.js"></script>
 </body>
 </html>
